Support enum member in enum member
