+++ packages/typespec-autorest/test/openapi-output.test.ts
@@ -672,137 +672,6 @@ describe("typespec-autorest: request", () => {
   });
 });
 
-describe("typespec-autorest: enums", () => {
-  const enumBase = Object.freeze({
-    type: "string",
-    enum: ["foo", "bar"],
-  });
-
-  it("create basic enum without values", async () => {
-    const res: any = await oapiForModel("Foo", `enum Foo {foo, bar}`);
-
-    const schema = res.defs.Foo;
-    deepStrictEqual(schema, {
-      ...enumBase,
-      "x-ms-enum": {
-        modelAsString: false,
-        name: "Foo",
-      },
-    });
-  });
-
-  it("enums are marked with modelAsString false by default", async () => {
-    const res: any = await oapiForModel("Foo", `enum Foo {foo, bar}`);
-
-    const schema = res.defs.Foo;
-    deepStrictEqual(schema["x-ms-enum"].modelAsString, false);
-  });
-
-  it("create enum with doc on a member", async () => {
-    const res: any = await oapiForModel(
-      "Foo",
-      `
-        enum Foo {
-          @doc("Foo doc")
-          foo,
-          bar
-        }
-      `,
-    );
-
-    const schema = res.defs.Foo;
-    deepStrictEqual(schema, {
-      ...enumBase,
-      "x-ms-enum": {
-        modelAsString: false,
-        name: "Foo",
-        values: [
-          {
-            description: "Foo doc",
-            name: "foo",
-            value: "foo",
-          },
-          {
-            name: "bar",
-            value: "bar",
-          },
-        ],
-      },
-    });
-  });
-
-  it("create enum with custom name/value on a member", async () => {
-    const res: any = await oapiForModel(
-      "Foo",
-      `
-        enum Foo {
-          FooCustom: "foo",
-          bar,
-        }
-      `,
-    );
-
-    const schema = res.defs.Foo;
-    deepStrictEqual(schema, {
-      ...enumBase,
-      "x-ms-enum": {
-        modelAsString: false,
-        name: "Foo",
-        values: [
-          {
-            name: "FooCustom",
-            value: "foo",
-          },
-          {
-            name: "bar",
-            value: "bar",
-          },
-        ],
-      },
-    });
-  });
-
-  it("create enum with numeric values", async () => {
-    const res: any = await oapiForModel(
-      "Test",
-      `
-        enum Test {
-          Zero: 0,
-          One: 1,
-        }
-      `,
-    );
-
-    const schema = res.defs.Test;
-    deepStrictEqual(schema, {
-      type: "number",
-      enum: [0, 1],
-      "x-ms-enum": {
-        modelAsString: false,
-        name: "Test",
-        values: [
-          {
-            name: "Zero",
-            value: 0,
-          },
-          {
-            name: "One",
-            value: 1,
-          },
-        ],
-      },
-    });
-  });
-
-  it("overrides x-ms-enum.name with @clientName", async () => {
-    const res: any = await compileOpenAPI(`@clientName("RenamedFoo") enum Foo {foo, bar}`, {
-      preset: "azure",
-    });
-    const schema = res.definitions.RenamedFoo;
-    deepStrictEqual(schema["x-ms-enum"].name, "RenamedFoo");
-  });
-});
-
 describe("typespec-autorest: extension decorator", () => {
   it("adds an arbitrary extension to a model", async () => {
     const oapi: any = await compileOpenAPI(`
